server/pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "dalar_rate"
version = "1.0.0"
description = "exchange_rate_dolar_curency"
readme = {file = "README.md", content-type = "markdown"}
requires-python = ">=3.8"
license = {file = "LICENSE.md"}
classifiers = [
    "Programming Language :: Python",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Private :: Do Not Upload"
]

dependencies = [
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.2",
    "ruff==0.4.10"
]

[tool.setuptools.packages.find]
where = ["src"]